# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrgUnitCreate": {
        "properties": {
          "name": {
            "maxLength": 255,
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "OrgUnitCreate",
        "type": "object"
      },
      "OrgUnitRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "parent_id",
          "created_at",
          "updated_at"
        ],
        "title": "OrgUnitRead",
        "type": "object"
      },
      "OrgUnitUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to leave unchanged; send null to make a root node",
            "title": "Parent Id"
          }
        },
        "title": "OrgUnitUpdate",
        "type": "object"
      },
...
      "Page_ScoreSnapshotRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScoreSnapshotRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ScoreSnapshotRead]",
        "type": "object"
      },
...
      "ScoreSnapshotRead": {
        "properties": {
          "app_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Id"
          },
          "avg_violations_per_page": {
            "title": "Avg Violations Per Page",
            "type": "number"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "org_unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org Unit Id"
          },
          "pages_with_critical_violations": {
            "title": "Pages With Critical Violations",
            "type": "integer"
          },
          "pages_with_violations": {
            "title": "Pages With Violations",
            "type": "integer"
          },
          "pct_pages_with_critical_violations": {
            "title": "Pct Pages With Critical Violations",
            "type": "number"
          },
          "pct_pages_with_violations": {
            "title": "Pct Pages With Violations",
            "type": "number"
          },
          "scan_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Run Id"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "snapshot_at": {
            "format": "date-time",
            "title": "Snapshot At",
            "type": "string"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          },
          "total_violations": {
            "title": "Total Violations",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "score",
          "total_violations",
          "pages_with_violations",
          "pages_with_critical_violations",
          "total_pages",
          "avg_violations_per_page",
          "pct_pages_with_violations",
          "pct_pages_with_critical_violations",
          "id",
          "app_id",
          "scan_run_id",
          "org_unit_id",
          "brand_id",
          "snapshot_at",
          "created_at",
          "updated_at"
        ],
        "title": "ScoreSnapshotRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Accessibility Health API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/org-units": {
      "get": {
        "operationId": "list_org_units",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrgUnitRead"
                  },
                  "title": "Response List Org Units",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Org Units",
        "tags": [
          "org-units"
        ]
      },
      "post": {
        "operationId": "create_org_unit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Org Unit",
        "tags": [
          "org-units"
        ]
      }
    },
    "/api/v1/org-units/{org_unit_id}": {
      "delete": {
        "operationId": "delete_org_unit",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Org Unit",
        "tags": [
          "org-units"
        ]
      },
      "get": {
        "operationId": "get_org_unit",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Org Unit",
        "tags": [
          "org-units"
        ]
      },
      "patch": {
        "operationId": "update_org_unit",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Org Unit",
        "tags": [
          "org-units"
        ]
      }
    },
    "/api/v1/org-units/{org_unit_id}/ancestors": {
      "get": {
        "operationId": "get_ancestors",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrgUnitRead"
                  },
                  "title": "Response Get Ancestors",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ancestors",
        "tags": [
          "org-units"
        ]
      }
    },
    "/api/v1/org-units/{org_unit_id}/descendants": {
      "get": {
        "operationId": "get_descendants",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrgUnitRead"
                  },
                  "title": "Response Get Descendants",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Descendants",
        "tags": [
          "org-units"
        ]
      }
    },
    "/api/v1/org-units/{org_unit_id}/scores": {
      "get": {
        "operationId": "list_org_unit_scores",
        "parameters": [
          {
            "in": "path",
            "name": "org_unit_id",
            "required": true,
            "schema": {
              "title": "Org Unit Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ScoreSnapshotRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Org Unit Scores",
        "tags": [
          "org-units"
        ]
      }
    },
...
  }
}